Fixed bug in sum as uint16_t was overflowing. Also increased range on no filter maximiser.

+++ FlashReceiver/features/Features.c
@@ -33,8 +33,8 @@ uint16_t Maximum(uint16_t size, uint16_t * array){
 	return(max);
 }
 
-uint16_t Sum(uint16_t size, uint16_t * array){
-	uint16_t retval = 0;
+uint64_t Sum(uint16_t size, uint16_t * array){
+	uint64_t retval = 0;
 	for(uint8_t i = 0; i < size; i++){
 		retval += array[i];
 	}
@@ -42,7 +42,7 @@ uint16_t Sum(uint16_t size, uint16_t * array){
 }
 
 uint16_t Average(uint16_t size, uint16_t * array){
-	uint32_t sum = Sum(size, array);
+	uint64_t sum = Sum(size, array);
 	return(sum/size);
 }
